update train/inference script

+++ cnn/layer/max_pooling.cpp
@@ -46,8 +46,8 @@ void MaxPooling::forward(const Matrix& bottom) {
 void MaxPooling::backward(const Matrix& bottom, const Matrix& grad_top) {
   grad_bottom.resize(bottom.rows(), bottom.cols());
   grad_bottom.setZero();
-  for (int i = 0; i < max_idxs.size(); i ++) {  // i-th sample
-    for (int j = 0; j < max_idxs[i].size(); j ++) {
+  for (size_t i = 0; i < max_idxs.size(); i++) {  // i-th sample
+    for (size_t j = 0; j < max_idxs[i].size(); j++) {
       grad_bottom(max_idxs[i][j], i) += grad_top(j, i);
     }
   }
